Remove dead code, synchronize with the last change.
Remove dead code in the Bt::CommandTask that was responsible for
managing bandwidth allocators on peer's sockets according to
currently selected allocation mode (limited/unlimited). Since
yesterday, bandwidth allocator became able to operate in the
unlimited mode so the managing code in the CommandTask class became
obsolete.

Synchronize with the last change in the Net::Socket class.